+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # unreleased
 * Add ability to/from import/export with TappedOut.net files
 * Add tag support
+* Move card preview to be inline in card list
 
 # 0.1.0
 * Add import and export buttons in settings modal